==> notebooks/0.data_preprocess.ipynb <==
import pandas as pd
from pathlib import Path
from molvs import standardize_smiles
from loguru import logger
from tqdm.auto import tqdm
from typing import Optional

def canonicalize_smiles(smiles: str) -> Optional[str]:
    try:
        return standardize_smiles(smiles)
    except Exception as e:
        logger.error(f"Error canonicalizing smiles: {smiles} with error: {e}")
        return None

raw_data_dir = Path("../datasets/raw")

data_consolidation = pd.read_csv(raw_data_dir / "Dataset_Consolidation.csv")
data_cyanine = pd.read_csv(raw_data_dir / "Dataset_Cyanine.csv")
data_xanthene = pd.read_csv(raw_data_dir / "Dataset_Xanthene.csv")

def canonicalize_df(df: pd.DataFrame, column: str) -> pd.DataFrame:
    tqdm.pandas(desc=f"canonicalizing {column}")
    logger.info(f"before canonicalizing # {column} in {df.shape[0]} rows")
    df[column] = df[column].progress_apply(canonicalize_smiles)
    df = df.dropna(subset=[column])
    logger.info(f"after canonicalizing # {column} in {df.shape[0]} rows")
    return df

data_consolidation = canonicalize_df(data_consolidation, "smiles")
data_consolidation = canonicalize_df(data_consolidation, "solvent")

data_cyanine = canonicalize_df(data_cyanine, "smiles")
data_cyanine = canonicalize_df(data_cyanine, "solvent")

data_xanthene = canonicalize_df(data_xanthene, "smiles")
data_xanthene = canonicalize_df(data_xanthene, "solvent")

data_consolidation.to_csv(raw_data_dir / "Dataset_Consolidation_canonicalized.csv", index=False)
data_cyanine.to_csv(raw_data_dir / "Dataset_Cyanine_canonicalized.csv", index=False)
data_xanthene.to_csv(raw_data_dir / "Dataset_Xanthene_canonicalized.csv", index=False)
