==> notebooks/1.split_datasets.ipynb <==
import pandas as pd
from pathlib import Path
from sklearn.model_selection import KFold
from loguru import logger
from tqdm.auto import tqdm
import random
from rdkit.Chem.Scaffolds import MurckoScaffold
from typing import List, Tuple
from typing import Optional, List

raw_data_dir = Path("../datasets/raw")

data_consolidation = pd.read_csv(raw_data_dir / "Dataset_Consolidation_canonicalized.csv")
data_cyanine = pd.read_csv(raw_data_dir / "Dataset_Cyanine_canonicalized.csv")
data_xanthene = pd.read_csv(raw_data_dir / "Dataset_Xanthene_canonicalized.csv")

def drop_duplicates(df: pd.DataFrame, columns: List[str]) -> pd.DataFrame:
    logger.info(f"before dropping duplicates: {df.shape[0]} rows")
    df = df.drop_duplicates(subset=columns)
    df.reset_index(drop=True, inplace=True)
    logger.info(f"after dropping duplicates: {df.shape[0]} rows")
    return df

data_consolidation = drop_duplicates(data_consolidation, ["smiles", "solvent"])
data_cyanine = drop_duplicates(data_cyanine, ["smiles", "solvent"])
data_xanthene = drop_duplicates(data_xanthene, ["smiles", "solvent"])

def random_split(
    df: pd.DataFrame, save_dir: Path, name: str, random_state: int = 42
) -> pd.DataFrame:
    kf = KFold(n_splits=5, shuffle=True, random_state=random_state)
    _save_dir = save_dir / "random"
    _save_dir.mkdir(parents=True, exist_ok=True)
    for fold, (train_index, valid_index) in enumerate(kf.split(df)):
        _df = df.copy()
        _df.loc[valid_index, "split"] = "valid"  # 这一折作为 valid
        _df.loc[train_index, "split"] = "train"  # 剩余的数据作为 train
        df_test = _df[_df["split"] == "valid"].copy()
        df_test["split"] = "test"
        _df = pd.concat([_df, df_test], ignore_index=True)
        _df.to_csv(_save_dir / f"{name}_fold{fold}.csv", index=False)

        n_total = len(_df)
        n_test = len(_df[_df["split"] == "test"])
        n_valid = len(_df[_df["split"] == "valid"])
        n_train = len(_df[_df["split"] == "train"])

        logger.info(
            f"length of {name}_fold{fold}: {n_total}; length of train: {n_train}; length of valid: {n_valid}; length of test: {n_test}"
        )

random_split(data_consolidation, raw_data_dir, "consolidation")
random_split(data_cyanine, raw_data_dir, "cyanine")
random_split(data_xanthene, raw_data_dir, "xanthene")

def generate_scaffold(smiles, include_chirality=False):
    """
    Obtain Bemis-Murcko scaffold from smiles
    :param smiles:
    :param include_chirality:
    :return: smiles of scaffold
    """
    scaffold = MurckoScaffold.MurckoScaffoldSmiles(
        smiles=smiles, includeChirality=include_chirality
    )
    return scaffold

def scaffold_split(
    smiles_list: List[str],
    k: int = 5,  # Number of folds
    balanced: bool = True,
    seed: int = 42,
) -> List[Tuple[List[int], List[int]]]:

    # Generate scaffold-based dictionary
    all_scaffolds = {}
    scaffolds = []
    for i, smiles in enumerate(tqdm(smiles_list)):
        try:
            scaffold = generate_scaffold(smiles, include_chirality=True)
            scaffolds.append(scaffold)
        except Exception:
            logger.warning(f"Error generating scaffold for {smiles}")
            continue
        if scaffold not in all_scaffolds:
            all_scaffolds[scaffold] = [i]
        else:
            all_scaffolds[scaffold].append(i)

    # Group scaffolds into a list of index sets
    scaffold_sets = list(all_scaffolds.values())

    # If balancing is enabled, shuffle larger scaffold sets differently
    if balanced:
        random.seed(seed)
        random.shuffle(scaffold_sets)

    # Create the KFold splits
    kf = KFold(n_splits=k, shuffle=True, random_state=seed)
    folds = []

    # We are going to assign each scaffold to one of the k folds
    for train_idx, val_idx in kf.split(scaffold_sets):
        train_fold = []
        val_fold = []

        # Collect indices for train and validation folds based on scaffolds
        for idx in train_idx:
            train_fold.extend(scaffold_sets[idx])
        for idx in val_idx:
            val_fold.extend(scaffold_sets[idx])

        folds.append((train_fold, val_fold))

    return folds, scaffolds

def scaffold_split_df(
    df: pd.DataFrame,
    name: str,
    k: int = 5,
    balanced: bool = True,
    seed: int = 42,
    save_dir: Path = raw_data_dir,
) -> List[Tuple[List[int], List[int]]]:
    smiles_list = df["smiles"].tolist()
    folds, scaffolds = scaffold_split(smiles_list, k=k, balanced=balanced, seed=seed)
    df["scaffold"] = scaffolds

    _save_dir = save_dir / "scaffold"
    _save_dir.mkdir(parents=True, exist_ok=True)

    n_scaffolds = df["scaffold"].nunique()
    n_smiles_unique = df["smiles"].nunique()
    logger.info(f"number of scaffolds: {n_scaffolds}")
    logger.info(f"number of smiles: {n_smiles_unique}")

    for fold, (train_idx, val_idx) in enumerate(folds):
        _df = df.copy()
        _df.loc[val_idx, "split"] = "valid"
        _df.loc[train_idx, "split"] = "train"
        df_test = _df[_df["split"] == "valid"].copy()
        df_test["split"] = "test"
        _df = pd.concat([_df, df_test], ignore_index=True)
        _df.to_csv(_save_dir / f"{name}_fold{fold}.csv", index=False)

        n_total = len(_df)
        n_test = len(_df[_df["split"] == "test"])
        n_valid = len(_df[_df["split"] == "valid"])
        n_train = len(_df[_df["split"] == "train"])

        logger.info(
            f"length of {name}_fold{fold}: {n_total}; length of train: {n_train}; length of valid: {n_valid}; length of test: {n_test}"
        )

scaffold_split_df(data_consolidation, "consolidation")

scaffold_split_df(data_cyanine, "cyanine")

scaffold_split_df(data_xanthene, "xanthene")
