==> notebooks/2.build_tasks.ipynb <==
import pandas as pd
from pathlib import Path
from loguru import logger

task_name = "absorption"  # absorption, emission, quantum_yield, log_molar_absorptivity
dataset_name = "xanthene"  # consolidation, cyanine, xanthene
split_method = "random"  # random, scaffold
raw_data_dir = Path("../datasets/raw")
data_dir = Path("../datasets/")

def build_data(raw_data_path, data_path, task_name):
    df = pd.read_csv(raw_data_path)
    logger.info(f"before dropna: {len(df)}")
    df.dropna(subset=[task_name], inplace=True)
    
    df_new = pd.DataFrame()
    df_new["smiles"] = df["smiles"]
    df_new["solvent"] = df["solvent"]
    df_new[task_name] = df[task_name]
    df_new["split"] = df["split"]

    n_total = len(df_new)
    n_test = len(df_new[df_new["split"] == "test"])
    n_valid = len(df_new[df_new["split"] == "valid"])
    n_train = len(df_new[df_new["split"] == "train"])
    df_new.to_csv(data_path, index=False)
    logger.info(f"Dataset {dataset_name} with {n_total} samples, {n_test} test, {n_valid} valid, {n_train} train")

for fold in range(5):
    logger.info(f"Building {task_name} for {dataset_name} fold {fold}")
    task_dir = data_dir / split_method / f"{dataset_name}_fold{fold}" / f"{task_name}"
    task_dir.mkdir(parents=True, exist_ok=True)

    raw_data_path = raw_data_dir / split_method / f"{dataset_name}_fold{fold}.csv"
    data_path = task_dir / f"{task_name}.csv"
    build_data(raw_data_path, data_path, task_name)
