==> notebooks/3.results.ipynb <==
import pandas as pd
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import numpy as np

df_valid = pd.read_csv(
    "/data/home/silong/projects/fluor/KPGT-Fluor/results/scaffold/cyanine_fold0/absorption/valid.csv"
)
df_train = pd.read_csv(
    "/data/home/silong/projects/fluor/KPGT-Fluor/results/scaffold/cyanine_fold0/absorption/train.csv"
)
df_results = pd.concat([df_train, df_valid], axis=0)
print(len(df_results))

r2_score(df_results["label"], df_results["predictions"])

mean_absolute_error(df_results["label"], df_results["predictions"])

np.sqrt(mean_squared_error(df_results["label"], df_results["predictions"]))
